move test to separate file under metrics build tag
